requirements-test.txt
# Requirements for testing
pytest==7.4.3
requests==2.31.0
mysql-connector-python==8.2.0
pytest-html==3.2.0
pytest-cov==4.1.0
